Refactor `Entry` with more separation from the `raw` part
